+++ test/gf/test_gf_psgrn_pscmp.py
@@ -7,6 +7,8 @@ from tempfile import mkdtemp
 import numpy as num
 import os
 import shutil
+from copy import deepcopy
+from multiprocessing import cpu_count
 
 from pyrocko import orthodrome as ortd
 from pyrocko import util, gf, cake  # noqa
@@ -61,20 +63,41 @@ def statics(engine, source, starget):
 class GFPsgrnPscmpTestCase(unittest.TestCase):
 
     tempdirs = []
+    pscmp_store_dir = None
 
     def __init__(self, *args, **kwargs):
-        self.pscmp_store_dir = None
         unittest.TestCase.__init__(self, *args, **kwargs)
+
+    @classmethod
+    def setUpClass(cls):
+        cls.pscmp_store_dir, cls.psgrn_config = \
+            cls._create_psgrn_pscmp_store(cls)
 
     @classmethod
     def tearDownClass(cls):
         for d in cls.tempdirs:
             shutil.rmtree(d)
 
+    def test_isolated_isotropic_component(self):
+        shear = 31126160000.0
+        lame_lambda = 25211680000.0
+
+        cdm = num.array([
+            [(2 * shear + lame_lambda), lame_lambda, lame_lambda],
+            [lame_lambda, (2 * shear + lame_lambda), lame_lambda],
+            [lame_lambda, lame_lambda, (2 * shear + lame_lambda)]])
+
+        nullf = psgrn_pscmp.get_nullification_factor(
+            mu=shear, lame_lambda=lame_lambda)
+
+        cdm[0:2, :] *= nullf
+        null_cdm = cdm.sum(0)
+
+        num.testing.assert_allclose(null_cdm[0:2], num.zeros(2))
+
     def get_pscmp_store_info(self):
         if self.pscmp_store_dir is None:
-            self.pscmp_store_dir, self.psgrn_config = \
-                self._create_psgrn_pscmp_store()
+            raise ValueError('Store does not exist!')
 
         return self.pscmp_store_dir, self.psgrn_config
 
@@ -100,6 +123,7 @@ mantle
  660. 10.79 5.965 4.229 1349. 549.6 5e17 1e19 1'''.lstrip()))
 
         store_dir = mkdtemp(prefix='gfstore')
+        # store_dir = '/tmp/pscmp_gfstore'
         self.tempdirs.append(store_dir)
         store_id = 'psgrn_pscmp_test'
 
@@ -124,52 +148,69 @@ mantle
             receiver_depth=0. * km,
             source_depth_min=0. * km,
             source_depth_max=15. * km,
-            source_depth_delta=.5 * km,
+            source_depth_delta=.25 * km,
             distance_min=0. * km,
             distance_max=50. * km,
-            distance_delta=.5 * km)
+            distance_delta=.25 * km)
 
         config.validate()
-        gf.store.Store.create_editables(
-            store_dir, config=config, extra={'psgrn_pscmp': c})
+        if not os.path.exists(os.path.join(store_dir, 'config')):
+            gf.store.Store.create_editables(
+                store_dir, config=config, extra={'psgrn_pscmp': c})
 
-        store = gf.store.Store(store_dir, 'r')
-        store.close()
+            store = gf.store.Store(store_dir, 'r')
+            store.close()
 
-        # build store
-        try:
-            psgrn_pscmp.build(store_dir, nworkers=1)
-        except psgrn_pscmp.PsCmpError as e:
-            if str(e).find('could not start psgrn/pscmp') != -1:
-                logger.warn('psgrn/pscmp not installed; '
-                            'skipping test_pyrocko_gf_vs_pscmp')
-                return
-            else:
-                raise e
+            # build store
+            try:
+                psgrn_pscmp.build(store_dir, nworkers=cpu_count())
+            except psgrn_pscmp.PsCmpError as e:
+                if str(e).find('could not start psgrn/pscmp') != -1:
+                    logger.warn('psgrn/pscmp not installed; '
+                                'skipping test_pyrocko_gf_vs_pscmp')
+                    return
+                else:
+                    raise e
 
         return store_dir, c
 
-    def test_fomosto_vs_psgrn_pscmp(self):
+    def fomosto_vs_psgrn_pscmp(self, pscmp_sources, gf_sources, atol=2*mm):
+
+        def plot_components_compare(fomosto_comps, psgrn_comps):
+            import matplotlib.pyplot as plt
+
+            fig, axes = plt.subplots(4, 3)
+
+            for i, (fcomp, pscomp, cname) in enumerate(
+                    zip(fomosto_comps, psgrn_comps, ['N', 'E', 'D'])):
+                fdispl = fcomp.reshape(nnorth, neast)
+                pdispl = pscomp.reshape(nnorth, neast)
+                pcbound = num.max([num.abs(pdispl.min()), pdispl.max()])
+                # fcbound = num.max([num.abs(fdispl.min()), fdispl.max()])
+
+                axes[0, i].imshow(
+                    pdispl, cmap='seismic', vmin=-pcbound, vmax=pcbound)
+                axes[1, i].imshow(
+                    fdispl, cmap='seismic', vmin=-pcbound, vmax=pcbound)
+                diff = pdispl - fdispl
+                rdiff = pdispl / fdispl
+                axes[2, i].imshow(diff, cmap='seismic')
+                axes[3, i].imshow(rdiff, cmap='seismic')
+
+                axes[0, i].set_title('PSCMP %s' % cname)
+                axes[1, i].set_title('Fomosto %s' % cname)
+                axes[2, i].set_title(
+                    'abs diff min max %f, %f' % (diff.min(), diff.max()))
+                axes[3, i].set_title(
+                    'rel diff min max %f, %f' % (rdiff.min(), rdiff.max()))
+
+            plt.show()
+
         store_dir, c = self.get_pscmp_store_info()
 
         origin = gf.Location(
             lat=10.,
             lon=-15.)
-
-        # test GF store
-        TestRF = dict(
-            lat=origin.lat,
-            lon=origin.lon,
-            depth=2. * km,
-            width=0.2 * km,
-            length=0.5 * km,
-            rake=uniform(-90., 90.),
-            dip=uniform(0., 90.),
-            strike=uniform(0., 360.),
-            slip=uniform(1., 5.))
-
-        source_plain = gf.RectangularSource(**TestRF)
-        source_with_time = gf.RectangularSource(time=123.5, **TestRF)
 
         N, E = num.meshgrid(num.linspace(-20. * km, 20. * km, nnorth),
                             num.linspace(-20. * km, 20. * km, neast))
@@ -177,7 +218,6 @@ mantle
         # direct pscmp output
         lats, lons = ortd.ne_to_latlon(
             origin.lat, origin.lon, N.flatten(), E.flatten())
-        pscmp_sources = [psgrn_pscmp.PsCmpRectangularSource(**TestRF)]
 
         cc = c.pscmp_config
         cc.observation = psgrn_pscmp.PsCmpScatter(lats=lats, lons=lons)
@@ -218,18 +258,120 @@ mantle
 
         engine = gf.LocalEngine(store_dirs=[store_dir])
 
-        for source in [source_plain, source_with_time]:
+        for source in gf_sources:
             t0 = time()
             r = engine.process(source, [starget_nn, starget_ml])
             logger.info('pyrocko stacking time %f' % (time() - t0))
-            for static_result in r.static_results():
+            for i, static_result in enumerate(r.static_results()):
                 un_fomosto = static_result.result['displacement.n']
                 ue_fomosto = static_result.result['displacement.e']
                 ud_fomosto = static_result.result['displacement.d']
 
-                num.testing.assert_allclose(un_fomosto, un_pscmp, atol=1*mm)
-                num.testing.assert_allclose(ue_fomosto, ue_pscmp, atol=1*mm)
-                num.testing.assert_allclose(ud_fomosto, ud_pscmp, atol=1*mm)
+                if show_plot:
+                    fomosto_comps = [un_fomosto, ue_fomosto, ud_fomosto]
+                    psgrn_comps = [un_pscmp, ue_pscmp, ud_pscmp]
+                    plot_components_compare(fomosto_comps, psgrn_comps)
+
+                num.testing.assert_allclose(un_fomosto, un_pscmp, atol=atol)
+                num.testing.assert_allclose(ue_fomosto, ue_pscmp, atol=atol)
+                num.testing.assert_allclose(ud_fomosto, ud_pscmp, atol=atol)
+
+    def test_fomosto_vs_psgrn_pscmp_shear(self):
+
+        origin = gf.Location(
+            lat=10.,
+            lon=-15.)
+
+        # test GF store
+        TestRF = dict(
+            lat=origin.lat,
+            lon=origin.lon,
+            depth=2. * km,
+            width=0.2 * km,
+            length=7. * km,
+            rake=uniform(-90., 90.),
+            dip=uniform(0., 90.),
+            strike=uniform(0., 360.),
+            slip=uniform(1., 5.))
+
+        source_plain = gf.RectangularSource(**TestRF)
+        source_with_time = gf.RectangularSource(time=123.5, **TestRF)
+
+        gf_sources = [source_plain, source_with_time]
+        pscmp_sources = [psgrn_pscmp.PsCmpRectangularSource(**TestRF)]
+
+        self.fomosto_vs_psgrn_pscmp(
+            pscmp_sources=pscmp_sources, gf_sources=gf_sources, atol=5*mm)
+
+    def test_fomosto_vs_psgrn_pscmp_tensile(self):
+
+        origin = gf.Location(
+            lat=10.,
+            lon=-15.)
+
+        # test GF store
+        TestRF = dict(
+            lat=origin.lat,
+            lon=origin.lon,
+            depth=2. * km,
+            width=2. * km,
+            length=5. * km,
+            rake=uniform(-90., 90.),
+            dip=uniform(0., 90.),
+            strike=uniform(0., 360.))
+
+        slip = uniform(1., 2.)
+
+        for open_mode in [-1., 1.]:   # closing, opening
+
+            source_plain = gf.RectangularSource(**TestRF)
+            source_plain.update(slip=slip, opening_fraction=open_mode)
+
+            source_with_time = deepcopy(source_plain)
+            source_with_time.update(time=123.5)
+
+            gf_sources = [source_plain, source_with_time]
+
+            pscmp_sources = [psgrn_pscmp.PsCmpRectangularSource(
+                opening=slip * open_mode, slip=0., **TestRF)]
+
+            self.fomosto_vs_psgrn_pscmp(
+                pscmp_sources=pscmp_sources, gf_sources=gf_sources, atol=7*mm)
+
+    def test_fomosto_vs_psgrn_pscmp_tensile_shear(self):
+
+        origin = gf.Location(
+            lat=10.,
+            lon=-15.)
+
+        # test GF store
+        TestRF = dict(
+            lat=origin.lat,
+            lon=origin.lon,
+            depth=3. * km,
+            width=2. * km,
+            length=5. * km,
+            rake=uniform(-90., 90.),
+            dip=uniform(0., 90.),
+            strike=uniform(0., 360.))
+
+        opening_fraction = 0.4
+        slip = uniform(1., 5.)
+        opening = slip * opening_fraction
+        pscmp_slip = slip - opening
+
+        source_plain = gf.RectangularSource(**TestRF)
+        source_plain.update(slip=slip, opening_fraction=opening_fraction)
+
+        source_with_time = deepcopy(source_plain)
+        source_with_time.update(time=123.5)
+
+        gf_sources = [source_plain, source_with_time]
+        pscmp_sources = [psgrn_pscmp.PsCmpRectangularSource(
+            opening=opening, slip=pscmp_slip, **TestRF)]
+
+        self.fomosto_vs_psgrn_pscmp(
+            pscmp_sources=pscmp_sources, gf_sources=gf_sources, atol=7*mm)
 
     def plot_gf_distance_sampling(self):
         origin = gf.Location(
@@ -250,8 +392,8 @@ mantle
 
         source_plain = gf.RectangularSource(**TestRF)
 
-        N, E = num.meshgrid(num.linspace(-40. * km, 40. * km, nnorth),
-                            num.linspace(-40. * km, 40. * km, neast))
+        N, E = num.meshgrid(num.linspace(-20. * km, 20. * km, nnorth),
+                            num.linspace(-20. * km, 20. * km, neast))
 
         starget_ml = gf.StaticTarget(
             lats=num.full(N.size, origin.lat),
@@ -364,5 +506,5 @@ mantle
 
 
 if __name__ == '__main__':
-    util.setup_logging('test_gf_psgrn_pscmp', 'warning')
+    util.setup_logging('test_gf_psgrn_pscmp', 'info')
     unittest.main()